save function added
write_protobuf works and checks for preexisting file

Pulled some preprocessing out of write_protobuf

Moved some of the preprocessing out of the write_protobuf
function in io.rs into a new save function in load.rs.
We will like rename load.rs into preprocess.rs in the near
future. I would also still like to break the save function
down into smaller chunks.

+++ src/load.rs
@@ -1,9 +1,10 @@
 //! Operations to manage loaded protos.
 
-use std::fs::File;
-use std::path::Path;
+use io::write_protobuf;
 use protos::master::Game;
 use screen::Interfaceable;
+use std::fs::File;
+use std::path::Path;
 
 #[allow(unused_mut)]
 pub fn load<I: Interfaceable>(
@@ -24,13 +25,36 @@ pub fn new<I: Interfaceable>(
 
     let name = src.prompt();
     game.set_name(name.clone());
+    return write_protobuf(src, game);
+}
 
-    src.print(&format!(
-        "Hello {}, please enter the name of the new save file:",
-        &name
-    ));
-    let path = src.prompt();
+pub fn save<I: Interfaceable>(src: &mut I, game: &Game) -> File {
+    let mut path;
+    let file;
+    loop {
+        src.print(&format!(
+            "Hello {}, please enter the name of the new save file:",
+            game.get_name()
+        ));
 
-    File::create(&Path::new(&path)).unwrap();
-    return Ok(game);
+        path = src.prompt();
+
+        if Path::new(&path).exists() {
+            src.confirm();
+            match src.prompt().parse() {
+                Ok(1) => {
+                    file = File::create(&Path::new(&path)).unwrap();
+                    break;
+                }
+                _ => {
+                    src.print("Ok let's try again then...");
+                    continue;
+                }
+            };
+        } else {
+            file = File::create(&Path::new(&path)).unwrap();
+            break;
+        }
+    }
+    return file;
 }
